EQ-251 Moved attains specific health checks to the api/attains urls.
