+++ app/server/app/routes/attains.js
@@ -1,14 +1,31 @@
 import cors from 'cors';
 import express from 'express';
 import Excel from 'exceljs';
+import { readdirSync, statSync } from 'node:fs';
+import path, { resolve } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import { tableConfig } from '../config/tableConfig.js';
 import { getActiveSchema } from '../middleware.js';
 import { appendToWhere, knex } from '../utilities/database.js';
 import { log } from '../utilities/logger.js';
 import StreamingService from '../utilities/streamingService.js';
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const jsonPageSize = parseInt(process.env.JSON_PAGE_SIZE);
 const maxQuerySize = parseInt(process.env.MAX_QUERY_SIZE);
+
+const minDateTime = new Date(-8640000000000000);
+const maxDateTime = new Date(8640000000000000);
+
+let isLocal = false;
+let isDevelopment = false;
+let isStaging = false;
+
+if (process.env.NODE_ENV) {
+  isLocal = 'local' === process.env.NODE_ENV.toLowerCase();
+  isDevelopment = 'development' === process.env.NODE_ENV.toLowerCase();
+  isStaging = 'staging' === process.env.NODE_ENV.toLowerCase();
+}
 
 class DuplicateParameterException extends Error {
   constructor(parameter) {
@@ -516,6 +533,138 @@ async function queryColumnValues(profile, columns, params, schema) {
   return await query;
 }
 
+/**
+ * Checks if the etl for the database ran successfully.
+ * @param {express.Request} req
+ * @param {express.Response} res
+ */
+async function checkDatabaseHealth(req, res) {
+  try {
+    // check etl status in db
+    let query = knex
+      .withSchema('logging')
+      .from('etl_status')
+      .select('database')
+      .first();
+    let results = await query;
+    if (results.database === 'failed') {
+      res.status(200).json({ status: 'FAILED-DB' });
+      return;
+    }
+
+    // verify the latest entry in the schema table is active
+    query = knex
+      .withSchema('logging')
+      .from('etl_schemas')
+      .select('active', 'creation_date')
+      .orderBy('creation_date', 'desc')
+      .first();
+    results = await query;
+    if (!results.active) {
+      res.status(200).json({ status: 'FAILED-SCHEMA' });
+      return;
+    }
+
+    // verify database updated in the last week, with 1 hour buffer
+    const timeSinceLastUpdate =
+      (Date.now() - results.creation_date) / (1000 * 60 * 60);
+    if (timeSinceLastUpdate >= 169) {
+      res.status(200).json({ status: 'FAILED-TIME' });
+      return;
+    }
+
+    // verify a query can be ran against each table in the active db
+    for (const profile of Object.values(tableConfig)) {
+      query = knex
+        .withSchema(req.activeSchema)
+        .from(profile.tableName)
+        .select(profile.idColumn)
+        .limit(1)
+        .first();
+      results = await query;
+      if (!results[profile.idColumn]) {
+        res.status(200).json({ status: 'FAILED-QUERY' });
+        return;
+      }
+    }
+
+    // everything passed
+    res.status(200).json({ status: 'UP' });
+  } catch (err) {
+    res.status(500).send('Error!' + err);
+  }
+}
+
+/**
+ * Checks if the etl for the domain values ran successfully.
+ * @param {express.Request} req
+ * @param {express.Response} res
+ */
+async function checkDomainValuesHealth(req, res) {
+  try {
+    // check etl status in db
+    const query = knex
+      .withSchema('logging')
+      .from('etl_status')
+      .select('domain_values')
+      .first();
+    const results = await query;
+    if (results.domain_values === 'failed') {
+      res.status(200).json({ status: 'FAILED-DB' });
+      return;
+    }
+
+    // initialize timeSinceLastUpdate to the minimum time node allows
+    let timeSinceLastUpdate = minDateTime;
+
+    // verify file update date is within the last week
+    if (isLocal) {
+      const path = resolve(__dirname, `../content-etl/domainValues`);
+
+      // get hours since file last modified
+      const files = readdirSync(path);
+
+      let oldestModifiedDate = maxDateTime;
+      files.forEach((file) => {
+        const stats = statSync(`${path}/${file}`);
+        if (stats.mtime < oldestModifiedDate) oldestModifiedDate = stats.mtime;
+      });
+
+      timeSinceLastUpdate =
+        (Date.now() - oldestModifiedDate) / (1000 * 60 * 60);
+    } else {
+      // setup public s3 bucket
+      setAwsConfig();
+
+      const s3 = new AWS.S3({ apiVersion: '2006-03-01' });
+
+      // get a list of files in the directory
+      const data = await s3
+        .listObjects({
+          Bucket: process.env.CF_S3_PUB_BUCKET_ID,
+          Prefix: 'content-etl/domainValues',
+        })
+        .promise();
+
+      let oldestModifiedDate = maxDateTime;
+      data.Contents.forEach((file) => {
+        if (file.LastModified < oldestModifiedDate)
+          oldestModifiedDate = file.LastModified;
+      });
+
+      timeSinceLastUpdate =
+        (Date.now() - oldestModifiedDate) / (1000 * 60 * 60);
+    }
+
+    // check that domain values was updated in the last week and 1 hour
+    res
+      .status(200)
+      .json({ status: timeSinceLastUpdate >= 169 ? 'FAILED-FILE' : 'UP' });
+  } catch (err) {
+    res.status(500).send('Error!' + err);
+  }
+}
+
 export default function (app, basePath) {
   const router = express.Router();
 
@@ -526,6 +675,10 @@ export default function (app, basePath) {
   };
 
   Object.entries(tableConfig).forEach(([profileName, profile]) => {
+    // ****************************** //
+    // Public / CORS Enabled          //
+    // ****************************** //
+
     // create get requests
     router.get(`/${profileName}`, cors(corsOptions), function (req, res) {
       executeQuery(profile, req, res);
@@ -546,6 +699,10 @@ export default function (app, basePath) {
       },
     );
 
+    // ****************************** //
+    // Private / NOT CORS Enabled     //
+    // ****************************** //
+
     // get column domain values
     router.get('/:profile/values/:column', function (req, res) {
       executeValuesQuery(req, res);
@@ -553,6 +710,14 @@ export default function (app, basePath) {
 
     router.post('/:profile/values/:column', function (req, res) {
       executeValuesQuery(req, res);
+    });
+
+    router.get('/health/etlDatabase', function (req, res) {
+      checkDatabaseHealth(req, res);
+    });
+
+    router.get('/health/etlDomainValues', async function (req, res) {
+      checkDomainValuesHealth(req, res);
     });
   });
